refactor: move hooks to a new folder
