+++ front-end/src/pages/public/LandingPage/SearchSection.tsx
@@ -1,7 +1,7 @@
 import { Search } from "lucide-react";
 import { Input } from "../../../components/ui/input";
 import { useEffect, useState } from "react";
-import { useDebounce } from "@/lib/useDebounce";
+import { useDebounce } from "@/hooks/useDebounce";
 import { useQuery } from "@tanstack/react-query";
 import { RepositoryCard } from "../../../components/RespositoryCard";
 import { RepositoriesData, Repository } from "@/types/repository";
